apps/task/pyproject.toml
[project]
name = "holy-grail-task"
version = "0.1.0"
description = "Celery task service for Holy Grail"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.4.0",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "email-validator>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["email_service", "tasks", "enums", "models", "schemas", "services"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.1",
    "mypy>=1.14.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "coverage>=7.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    ".venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["."]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
